Move changes to new test suite
