+++ imagetest/test_suites/gveperf/startupscripts/netserver_startup.sh
@@ -10,14 +10,10 @@ sleepduration=600
 if [[ -f /usr/bin/apt ]]; then
   echo "$(date +"%Y-%m-%d %T"): apt found Installing iperf." | tee -a "$outfile"
   sudo apt update && sudo apt install iperf | tee -a "$outfile"
-fi
-
-if [[ -f /bin/apt-get ]]; then
+elif [[ -f /bin/apt-get ]]; then
   echo "$(date +"%Y-%m-%d %T"): apt found Installing iperf." | tee -a "$outfile"
   sudo apt-get update && sudo apt-get install iperf | tee -a "$outfile"
-fi
-
-if [[ -f /bin/dnf ]]; then
+elif [[ -f /bin/dnf ]]; then
   echo "$(date +"%Y-%m-%d %T"): dnf found Installing iperf." | tee -a "$outfile"
   os=$(cat /etc/redhat-release)
   if [[ "$os" == *"release 9."* ]]; then
@@ -29,15 +25,11 @@ if [[ -f /bin/dnf ]]; then
   sudo dnf -y install https://dl.fedoraproject.org/pub/epel/epel-release-latest-8.noarch.rpm | tee -a "$outfile"
   sudo dnf -y install https://dl.fedoraproject.org/pub/epel/epel-release-latest-9.noarch.rpm | tee -a "$outfile"
   sudo sudo dnf makecache && sudo dnf -y install iperf | tee -a "$outfile"
-fi
-
-if [[ ! -f /bin/dnf ]] && [[ -f /bin/yum ]]; then
+elif [[ ! -f /bin/dnf ]] && [[ -f /bin/yum ]]; then
   echo "$(date +"%Y-%m-%d %T"): yum found Installing iperf." | tee -a "$outfile"
   yum install https://dl.fedoraproject.org/pub/epel/epel-release-latest-7.noarch.rpm | tee -a "$outfile"
   sudo sudo yum makecache && sudo yum -y install iperf | tee -a "$outfile"
-fi
-
-if [[ -f /usr/bin/zypper ]]; then
+elif [[ -f /usr/bin/zypper ]]; then
   echo "$(date +"%Y-%m-%d %T"): zypper found Installing iperf." | tee -a "$outfile"
   sudo zypper --no-gpg-checks refresh | tee -a "$outfile"
   sudo zypper --no-gpg-checks --non-interactive install https://iperf.fr/download/opensuse/iperf-2.0.5-14.1.2.x86_64.rpm | tee -a "$outfile"
